requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
azure-core==1.26.3
azure-storage-blob==12.15.0
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1
django-celery-results==2.5.0
django-cloudinary-storage==0.3.0
django-configurations==2.2
django-cors-headers==3.13.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-extra-fields==3.0.2
django-hashids==0.7.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
firebase-admin==6.1.0
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
hashids==1.3.1
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==22.0
Pillow==9.4.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycryptodome==3.18.0
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
resolvelib==0.8.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
whitenoise==5.2.0
zope.interface==6.0
